https://leetcode.com/explore/interview/card/amazon/76/array-and-strings/2967/
All <, >, = follows the same approach but in case of closest, it gets little tricky.
To move the pointers, we still follow the old approach but to update the min, we will have to check the absolute distance and then update.
Why move pointers the old way?
If i and j are at beginning and at end, and sum > K, moving the value of i will further increase the distance, so moving i is of no use, only moving j might help. So move j. Likewise in case of sum < K. But to know the closest distance just check the absolute value.