https://leetcode.com/explore/interview/card/amazon/79/sorting-and-searching/2991/
https://www.youtube.com/watch?v=LPFhl65R7ww
The idea is to get the shortest of the 2 inputs and then move in that using the Binary Search technique. The four conditions along the cut are important to understand.
If maxLeft1 > minRight2, it means that the left array has a bigger element, so moving the low to right will not help as it will still produce the same condition, so move it left. 

Idea is to move all the greatest element to the RIGHT and keep all the smallest to the LEFT.